OrderItemFile renamed to File and moved. SyncVariantFileResponse replaced by File
